Fix bad namespace rename
