+++ src/app/Http/Controllers/Api/GameWarningController.php
@@ -8,7 +8,7 @@ use App\Services\PlayerWarnings\PlayerWarningService;
 use App\Services\PlayerBans\ServerKeyAuthService;
 use App\Services\PlayerBans\Exceptions\UnauthorisedKeyActionException;
 use App\Entities\Eloquent\ServerKeys\Models\ServerKey;
-use App\Entities\Eloquent\GameIdentifierType;
+use App\Entities\GameIdentifierType;
 
 final class GameWarningController extends ApiController
 {
