+++ src/app/Services/Queries/ServerQueryService.php
@@ -7,7 +7,7 @@ use App\Services\Queries\Jobs\ServerQueryJob;
 use App\Services\Queries\Jobs\PlayerQueryJob;
 use App\Services\Queries\Entities\ServerJobEntity;
 use App\Entities\Eloquent\Servers\Repositories\ServerStatusPlayerRepository;
-use App\Entities\Eloquent\GameType;
+use App\Entities\GameType;
 
 final class ServerQueryService
 {
