+++ src/routes/web.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Entities\Eloquent\Environment;
+use App\Entities\Environment;
 
 /*
 |--------------------------------------------------------------------------
